Handling existing files

+++ src/main/java/com/eshop/services/implementation/FileHandlerServiceImpl.java
@@ -16,7 +16,11 @@ public class FileHandlerServiceImpl implements FileHandlerService {
     @Override
     public String handleFileUpload(MultipartFile file, Path path) {
         try {
-            Files.copy(file.getInputStream(), path.resolve(file.getOriginalFilename()));
+            if(!Files.exists(path.resolve(file.getOriginalFilename()))) {
+                Files.copy(file.getInputStream(), path.resolve(file.getOriginalFilename()));
+            } else{
+                return "alreadyExists";
+            }
         } catch (IOException e) {
             return "error";
         }
